mavlink: move GROUND_TRUTH stream to separate file
